requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
imbalanced-learn
scipy
shap
kagglehub
tensorflow
pillow
opencv-python
pydot
sklearn-genetic-opt
google-genai
fastapi
uvicorn[standard]
streamlit
python-dotenv
